frontend/requirements-dev.txt
# Test dependencies
pytest==7.4.3
pytest-mock==3.12.0
requests-mock==1.11.0
